storage: move new-style persist_sink code to storage, use same code for both compute and storage
